+++ framework/changelog.md
@@ -1,5 +1,6 @@
 <!-- The pattern we follow here is to keep the changelog for the latest version -->
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
-- chore: Upgrades core to 1.2.19
+- chore: Upgrades core to 1.2.20
 - feat: add selected key and virtual key to logs table
+- feat: add headers to MCP client config